pyproject.toml
[project]
name = "hans-under-exit"
version = "0.1.0"
description = ""
authors = [
    {name = "Danylo",email = "dchynonyi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "datetime (>=6.0,<7.0)",
    "vectorbt (>=0.28.2,<0.29.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
tqdm = "^4.67.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
